Add files via upload
The Computer Science Department follows certain criteria when a student learns to program. A
number of programming exercises must be worked through. To proceed to the next exercise a student has
to obtain a mark of 50% or more and must have completed 5 or more program runs. You are requested
to write a program to validate if a student can proceed to the next program.